@@ -4,7 +4,6 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 exports.default = createCodeBlockWithLineNumbers;
-var _react = _interopRequireDefault(require("react"));
 var _reactInnertext = _interopRequireDefault(require("react-innertext"));
 function _interopRequireDefault(e) { return e && e.__esModule ? e : { default: e }; }
 function createCodeBlockWithLineNumbers(OrigPre) {
@@ -13,12 +12,14 @@ function createCodeBlockWithLineNumbers(OrigPre) {
     numberOfLines,
     ..._props
   }) => {
-    return /*#__PURE__*/_react.default.createElement("code", {
+    return /*#__PURE__*/React.createElement("code", {
       className: "line-numbers-of-code-block"
     }, Array.from({
       length: numberOfLines
     }, (_, i) => {
-      return /*#__PURE__*/_react.default.createElement("span", null, `${(i + 1).toString().padStart(3)}\n`);
+      return /*#__PURE__*/React.createElement("span", {
+        key: i
+      }, `${(i + 1).toString().padStart(3)}\n`);
     }));
   };
   // Code block with line numbers
@@ -31,16 +32,16 @@ function createCodeBlockWithLineNumbers(OrigPre) {
       children,
       ...props
     }) => {
-      return OrigPre ? /*#__PURE__*/_react.default.createElement(OrigPre, props, children) : /*#__PURE__*/_react.default.createElement("pre", props, children);
+      return OrigPre ? /*#__PURE__*/React.createElement(OrigPre, props, children) : /*#__PURE__*/React.createElement("pre", props, children);
     };
     // Render
-    return /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, /*#__PURE__*/_react.default.createElement(OrigNode, props, /*#__PURE__*/_react.default.createElement("div", {
+    return /*#__PURE__*/React.createElement(React.Fragment, null, /*#__PURE__*/React.createElement(OrigNode, props, /*#__PURE__*/React.createElement("div", {
       className: "code-block-line-number__container"
-    }, /*#__PURE__*/_react.default.createElement("div", {
+    }, /*#__PURE__*/React.createElement("div", {
       className: "code-block-line-number__code-contents"
-    }, children), /*#__PURE__*/_react.default.createElement("div", {
+    }, children), /*#__PURE__*/React.createElement("div", {
       className: "code-block-line-number__line-numbers"
-    }, /*#__PURE__*/_react.default.createElement(LineNumbersOfCodeBlock, {
+    }, /*#__PURE__*/React.createElement(LineNumbersOfCodeBlock, {
       numberOfLines: (0, _reactInnertext.default)(children).split(/\n/).length - 1
     })))));
   };
